pyproject.toml
[project]
name = "global-agent"
version = "0.1.0"
description = "Um projeto focado na automação total por IA"
authors = [
    {name = "Juan",email = "juan.bisposantos05@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai (>=0.120.1,<0.121.0)",
    "crewai-tools (>=0.45.0,<0.46.0)",
    "qdrant-client (>=1.14.2,<2.0.0)",
    "langchain-qdrant (>=0.2.0,<0.3.0)",
    "langchain-huggingface (>=0.2.0,<0.3.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "sentence-transformers (>=4.1.0,<5.0.0)",
    "embedchain (>=0.1.128,<0.2.0)"
]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"